added an AX disasm with (few) comments. Moved some code around to prepare to share some breakpoint code between the ppc and dsp
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3566 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_Wiimote/Src/wiimote_real.cpp
@@ -147,7 +147,7 @@ void ReadData()
     {
 		//INFO_LOG(CONSOLE, "Writing data to the Wiimote\n");
         SEvent& rEvent = m_EventWriteQueue.front();
-		wiiuse_io_write(m_pWiiMote, (byte*)rEvent.m_PayLoad, m_EventWriteQueue.size());
+		wiiuse_io_write(m_pWiiMote, (byte*)rEvent.m_PayLoad, (int)m_EventWriteQueue.size());
 		m_EventWriteQueue.pop();
 		
 #ifdef _WIN32
